progressbar: don't display progress bar if disabled
When disabling the progress bar don't display it if it is already
disabled.  This allows code to simply call disable_progress_bar() rather
than having to first check if it is enabled.

Also if the progress_bar is enabled, print a newline afterwards, this is
to ensure the next message is on the next line.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>