mksquashfs: make EXIT_MKSQUASHFS() etc restore via the restore thread
Since parallelisation only the main thread has been able to
call the EXIT_MKSQUASHFS macros, the sub-threads have not been
able to because the exit macros used to do a longjmp on restore
which is only valid on the main thread.

Unfortunately there are instances where the sub-threads do execute
code which call the exit macros, these are usually (but not exclusively)
in code which is mostly called by the main thread.  If these exit macros
ever triggered on the sub-threads then it wouldn't have worked.
Removing these exit macros was by and large difficult because they are
needed for the main thread.

But now there exists a restore thread, it is easy to change the exit
macros to signal the restore thread and then to exit the calling thread.
This works for the main thread and the sub-threads, meaning the exit
macros can now be used anywhere and by any thread.  This removes the
constant need when writing code to check what threads will be calling it,
and of course it fixes the existing bugs.

The other upshot of this change is all restoring is now done via the
restore thread.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>