mksquashfs: simplify pthread_cleanup_pop handling
If exiting via EXIT_MKSQUASHFS(), BAD_ERROR() or MEM_ERROR()
we do not need to explicitly execute the pthread_cleanup_pop
handler because these macros terminate the thread via
pthread_exit, which executes all pthread_cleanup handlers itself.
If we're not restoring, we simply exit the process, which is fine.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>