queues-caches-lists: move cache freelist allocate logic
into here, and make it per a per cache option.

On appending we want to change the allocate logic for the
fragment and writer caches to be grow first, rather than allocate
from the free list first.  This is because on appending with the use
free list first logic the caches never grow very large leading
to reduced cache effectiveness.

By making it a per cache option we fix the issue that on appending
we also made the reader cache grow first.  This was *not* a major
problem except because we never do lookup on the reader cache we
don't need to do it!

This commit is mainly because it always should have been a
cache private variable rather than a global.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>