caches-queues-lists: get rid of "keep" blocks and make behaviour more explicit
Originally the caches in mksquashfs were "buffer pools" that
expanded as necessary when readahead/number of buffers in flight
increased, and shrank as the number of buffers in flight decreased.

It was realised that the "buffer pools" could also be used to lookup
blocks in duplicate checking.  Firstly, when the fragment/blocks of the
file being duplicate checked were "still in flight" it was faster to
look them up in the "buffer pool" rather than have to wait for the
blocks to be written to disk (as was originally necessary because
blocks in flight were inaccessible).  Secondly, when blocks were
given back to the "buffer pool", if rather than shrinking the
cache, the blocks were put onto a freelist, duplicate checking might
find the blocks there which is faster than having to read them back
from the output filesystem.

From this evolved the concept of "keep" blocks, which when a
block was obtained from the cache via cache_get() if the keep flag
was set this signified that the block was to be kept when released
via cache_put().

This was perfectly adequate until now when code is being added to
monitor/record cache behaviour.  As the keep flag is a per block
flag there is no cache specific information that indicates whether
a cache is being used as a non-shrinking lookup cache or as shrinking
non-lookup based "cache" (in this case it is really acting as a buffer
pool).  This is important because it turns out the statistics to be
gathered differs depending on the cache usage.  A non-shrinking lookup
cache has count which represents the maximum size the cache has
grown to, and used which represents the size of the cache currently in
use.  A shrinking non-lookup cache on the other hand shrinks as blocks
become unused, and so count and used are one and the same.  Given this
there is no way of determining what the maximum size the shrinking
cache has historically grown to using just count and used; used is
completely redundant, and the maximum historical size is not recorded.

So as first step, rearchitect the "keep" code moving the flag from
being a per block flag to being a per cache flag, making it explicit
whether the cache is being used as a non-shrinking lookup cache or as
a shrinking non-lookup based cache.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>