unsquashfs: make {disable|enable}_progress_bar contingent on progress flag
By doing this move logic of whether the progress bar is enabled
inside the functions, allowing {disable|enable}_progress_bar to
be called multiply to disable and enable the progress bar without
the caller knowing whether the progress bar is enabled or not.

Previously these functions unconditionally enabled and disabled
the progress bar with the associated logic in the caller.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>