unsquashfs: Call compressor_extract_options
Unsquashfs should have been calling compressor_extract_options to
check that it knew how to decompress a filesystem compressed with
the options stored in the filesystem.

As it happens xz compression up till now has been the only
compressor to use compression options, and the code has always
been able to decompress xz compressed filesystems whatever the
compression options used, and so the check hasn't been needed.  But this
may change with other compressors.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>