xz_wrapper: make global comp_opts a static inside xz_dump_options
This was global purely to allow xz_dump_options() to return
a pointer to it.  However, making it global runs the risk coding
mistakes which should refer to a local comp_opts are hidden as they
pick up the global instead.  As the global is potentially swapped and
not in host byte order this isn't a good thing.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>