lzo_wrapper: Remove current implementation!
There has been numerous complaints against the LZO implementation,
namely that it takes longer to compress than GZIP, but produces
worse compression.  As such for mksquashfs, it makes no sense
to use LZO compression to reduce build time when producing
experimental images.

The latest "complaint" is:
http://sourceforge.net/p/squashfs/support-requests/3

I produced an implementation of LZ4 for Squashfs, which as LZ4
is extremely fast at compression, I intended this to be used
as an extremely fast build time option for building experimental
images.  But, I have received unexpected resistance to adding
LZ4 support to Squashfs in the kernel, and so it's basically
back to the drawing board.

I have decided to re-write the LZO implementation, adding
support for different LZO compression algorithms, with the
addition of compressor command line options to select which
algorithm is used, together with all the other bits which
need to be added to support the "compression option" framework.

Note the original implementators of the LZO support can't be blamed for
not adding this support, because the "compression options" framework
didn't exist when the implementation was made!

The current LZO wrapper has an implemention for the lzo1x_999_compress
algorithm, this has to be completely rewritten to deal gracefully with the
ability to deal with multiple compression algorithms (a table of the
different compressors, and function pointers to the appropriate
function), and so the current implementation is not salvageable.

The LZ4 implementation incorporates the necessary support for the
compression options framework, and this code can be reused in the new
LZO wrapper, recouping some of the "wasted" effort in its implementation.

So, on the one-hand the current LZO implementation lacks salvageable code, and
on the other hand, the LZ4 implementation incorporates a compression
options framework which can be reused in the new LZO wrapper.

So, it makes sense to remove the current implementation, and base the
new implementation on the LZ4 wrapper.

Signed-off-by: Phillip Lougher <phillip@squashfs.rg.uk>