Mksquashfs: move progress bar disable
Originally when file writing had finished in dir_scan() the
progress bar was disabled.  However, at this point there can
be queued fragments still to be compressed, and blocks queued
still to be written.

On a fast system and fast output device, this doesn't matter, as
the gap between the progress bar stopping and mksquashfs finishing
is small, as any fragments and blocks are quickly dealt with.

However, on a slow system or a slow output device, large queues
of fragments and/or blocks can have built up, which may time a while
to compress and/or write.  Having the progress bar stop before
this is done can be confusing, as it looks like Mksquashfs has hung.

So move progress bar disabling to after fragment compression and
block writing has finished.

Signed-off-by: Phillip Lougher <phillip@squashfs.org.uk>